Enemy Stats And Health

+++ Assets/_Scripts/Weapon/WeaponStats.cs
@@ -12,7 +12,7 @@ public class WeaponStats : ScriptableObject
     public float range; // Tầm bắn
     public float fireRate; // Tốc độ bắn (số phát bắn mỗi giây)
     public float detectionRange; // Tầm phát hiện enemy
-    public float damage; // Sát thương
+    public int damage; // Sát thương
     public float CritChance;
     public float CritDamage;
     public float fireCooldown;
